Prune more unnecessary includes from CUDA transformers
These kernels are incredibly slow to compile and for the most part are
completely independant of ATen/c10 yet they still end up including
half of `c10` transitively through `CUDAGeneratorImpl.h` and
`CUDAContext.h`.

This trims the fat so `mem_eff_attention` doesn't depend on ATen/c10 at all,
and `flash_attn` now only depends on `PhiloxUtils.cuh` (split out from
`CUDAGeneratorImpl.h`) and `CUDAContextLight.h` which doesn't transitively
include `TensorImpl.h`.

[ghstack-poisoned]
