+++ aten/src/ATen/cuda/CUDAGraphsUtils.cuh
@@ -2,7 +2,7 @@
 
 #include <ATen/cuda/CUDAGeneratorImpl.h>
 #include <ATen/cuda/CUDAEvent.h>
-#include <ATen/cuda/detail/UnpackRaw.cuh>
+#include <ATen/cuda/PhiloxUtils.cuh>
 #include <ATen/cuda/detail/CUDAHooks.h>
 #include <ATen/detail/CUDAHooksInterface.h>
 #include <c10/core/StreamGuard.h>
