+++ aten/src/ATen/cuda/PhiloxCudaState.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <cstdint>
+
+#include <ATen/cuda/detail/PhiloxCudaStateRaw.cuh>
