+++ aten/src/ATen/cuda/PhiloxUtils.cuh
@@ -1,0 +1,4 @@
+#pragma once
+
+#include <ATen/cuda/PhiloxCudaState.h>
+#include <ATen/cuda/detail/UnpackRaw.cuh>
