+++ aten/src/ATen/native/transformers/cuda/flash_attn/flash_bwd_kernel.h
@@ -4,6 +4,8 @@
 
 #pragma once
 
+#include <ATen/cuda/PhiloxUtils.cuh>
+
 #include <cute/algorithm/copy.hpp>
 #include <cute/algorithm/gemm.hpp>
