+++ aten/src/ATen/native/transformers/cuda/flash_attn/kernel_traits.h
@@ -4,8 +4,6 @@
 
 #pragma once
 
-#include <ATen/cuda/CUDAContext.h>
-
 #include <cute/algorithm/copy.hpp>
 
 #include <cutlass/cutlass.h>
