+++ aten/src/ATen/native/transformers/cuda/mem_eff_attention/epilogue/epilogue_pipelined.h
@@ -50,8 +50,6 @@
 #include <cassert>
 #endif
 
-#include <ATen/cuda/CUDAContext.h>
-
 #include <cutlass/aligned_buffer.h>
 #include <cutlass/array.h>
 #include <cutlass/cutlass.h>
