+++ aten/src/ATen/native/transformers/cuda/mem_eff_attention/epilogue/epilogue_rescale_output.h
@@ -18,8 +18,6 @@
 #include <cassert>
 #endif
 
-#include <ATen/cuda/CUDAContext.h>
-
 #include <cutlass/aligned_buffer.h>
 #include <cutlass/array.h>
 #include <cutlass/cutlass.h>
