+++ aten/src/ATen/native/transformers/cuda/mem_eff_attention/epilogue/epilogue_thread_apply_logsumexp.h
@@ -35,8 +35,6 @@
 
 #pragma once
 
-#include <ATen/cuda/CUDAContext.h>
-
 #include <cuda_fp16.h>
 
 #include <cutlass/array.h>
